requirements.txt
googlesearch-python
requests
bs4
langchain_experimental==0.0.60